requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.7
colorlog==6.7.0
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
deepspeed==0.12.6
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
entrypoints==0.4
fasteners==0.18
ffmpeg==1.4
filelock==3.12.4
flamingo-pytorch==0.1.2
fonttools==4.44.0
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.10.0
huggingface-hub==0.18.0
hydra-colorlog==1.2.0
hydra-core==1.1.1
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.1.1
iopath==0.1.10
joblib==1.2.0
kiwisolver==1.4.5
lightning-lite==1.8.6
lightning-utilities==0.9.0
markdown-it-py==2.2.0
matplotlib==3.7.3
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numcodecs==0.12.1
numpy==1.24.4
numpy-quaternion==2022.4.3
nvidia-ml-py==12.535.133
nvitop==1.3.1
oauthlib==3.2.2
omegaconf==2.1.2
open-clip-torch==2.20.0
# -e git+https://github.com/RoboFlamingo/RoboFlamingo.git@e5cde1c24612fdaa162d50fd1cb593f907f26943#egg=open_flamingo&subdirectory=open_flamingo
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==9.5.0
plotly==5.18.0
portalocker==2.8.2
proglog==0.1.10
protobuf==3.20.3
psutil==5.9.6
pyasn1==0.5.0
pyglet==2.0.10
Pygments==2.16.1
pyhash==0.9.3
PyOpenGL==3.1.0
pyparsing==3.1.1
pyrender==0.1.45
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytorch3d==0.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.5
rsa==4.9
sacremoses==0.0.53
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.5.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.10
tokenizers==0.13.3
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.33.1
trimesh==4.0.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wandb==0.15.2
wcwidth==0.2.9
webdataset==0.2.75
yacs==0.1.8
yarl==1.9.2
zarr==2.15.0
zipp==3.17.0
